NSIS use readthedocs PDFs

+++ installers/win32/create_nsis.sh
@@ -11,6 +11,19 @@ run_script() {
 
     cp ../../installers/win32/* ./
 
+    if ! curl -o OoliteReadMe.pdf -L https://oolite.readthedocs.io/en/latest/index.pdf; then
+       echo "❌ OoliteReadMe.pdf (index.pdf) download failed!" >&2
+       return 1
+    fi
+    if ! curl -o AdviceForNewCommanders.pdf -L https://oolite.readthedocs.io/en/latest/advice/advice.pdf; then
+       echo "❌ AdviceForNewCommanders.pdf (advice.pdf) download failed!" >&2
+       return 1
+    fi
+    if ! curl -o OoliteRS.pdf -L https://oolite.readthedocs.io/en/latest/OoliteRS.pdf; then
+       echo "❌ OoliteRS.pdf download failed!" >&2
+       return 1
+    fi
+
     # Passing arguments cause problems with some versions of NSIS.
     # Because of this, we generate them into a separate file and include them.
     echo "; Version Definitions for Oolite" > OoliteVersions.nsh
